BB: Rename noop_test_configuration.py -> noop.py
